+++ src/index.ts
@@ -1,3 +1,3 @@
-import ChromeWebStoreAPI from './chrome-web-store-api';
-export default ChromeWebStoreAPI;
+import ChromeWebStore from "./chrome-web-store";
+export default ChromeWebStore;
 export * from './chrome-web-store-api';
