+++ test/chrome-web-store-api.spec.ts
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
-import ChromeWebStoreAPI from '../src/chrome-web-store-api';
+import ChromeWebStore from '../src/chrome-web-store';
 
 describe('ChromeWebStoreAPI', () => {
 	before(async function () {
-		this.chromeWebStoreAPI = new ChromeWebStoreAPI(
+		this.chromeWebStoreAPI = new ChromeWebStore(
 			JSON.parse(process.env.CHROME_WEB_STORE_API_CREDENTIAL || ''),
 			JSON.parse(process.env.CHROME_WEB_STORE_API_ACCESS_TOKEN_RESPONSE || ''),
 		);
@@ -11,7 +11,7 @@ describe('ChromeWebStoreAPI', () => {
 
   describe('Item', () => {
 		it('should return Item Resource', async function() {
-			const chromeWebStoreAPI = this.chromeWebStoreAPI as ChromeWebStoreAPI;
+			const chromeWebStoreAPI = this.chromeWebStoreAPI as ChromeWebStore;
 			const itemId = 'pgpnkghddnfoopjapnlklllpjknnibkn';
 			const item = await (new chromeWebStoreAPI.Item(itemId)).fetch();
 			expect(item).to.have.property('id', itemId);
@@ -20,7 +20,7 @@ describe('ChromeWebStoreAPI', () => {
 
 	describe.skip('InAppProduct', () => {
     it('should return InAppProduct Resource', async function() {
-			const chromeWebStoreAPI = this.chromeWebStoreAPI as ChromeWebStoreAPI;
+			const chromeWebStoreAPI = this.chromeWebStoreAPI as ChromeWebStore;
       const itemId = 'pgpnkghddnfoopjapnlklllpjknnibkn';
       const item = new chromeWebStoreAPI.Item(itemId);
       const sku = '';
